Split tests into files: Crowdsale, DAO, LifToken
